+++ app/layout.tsx
@@ -2,7 +2,6 @@
 import "@/styles/globals.css";
 import { Inter } from "@next/font/google";
 import Toaster from "@/components/toaster";
-import AuthStatus from "@/components/auth-status";
 
 const inter = Inter({
   variable: "--font-inter",
@@ -13,12 +12,10 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const AuthStatusDiv = await AuthStatus();
   return (
     <html lang="en">
       <body className={inter.variable}>
         <Toaster />
-        {AuthStatusDiv}
         {children}
       </body>
     </html>
